+++ tests/tuxemon/test_capture.py
@@ -9,7 +9,8 @@ from tuxemon.formula import (
     capture,
     shake_check,
 )
-from tuxemon.monster import Monster, MonsterStatusHandler
+from tuxemon.monster import Monster
+from tuxemon.monster_dir.status import MonsterStatusHandler
 
 
 class TestShakeCheck(unittest.TestCase):
